producer/requirements.txt
annotated-types==0.7.0
certifi==2025.8.3
charset-normalizer==3.4.3
dnspython==2.7.0
idna==3.10
kafka-python==2.2.15
pydantic==2.11.7
pydantic_core==2.33.2
pymongo==4.13.2
python-dotenv==1.1.1
requests==2.32.5
typing-inspection==0.4.1
typing_extensions==4.15.0
urllib3==2.5.0
